% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{departure_stop}
\alias{departure_stop}
\title{Number of departures per stop}
\usage{
departure_stop(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{GTFS object loaded with tidytransit::read_gtfs()}
}
\value{
dep Table with number of departures per stop ID
}
\description{
Create dataframe with the number of departures per stop. Ensure to filter GTFS feed by date (tidytransit::filter_feed_by_date())
prior to running the function to get number of departures per day
}
\examples{
\donttest{
library(dplyr)
gtfs_file <- list.files("gtfs_data")
dat = tidytransit::read_gtfs(paste0("gtfs_data/", gtfs_file[1]))
departure_stop(dat)
}

}
