% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{departure_stop_line_hr}
\alias{departure_stop_line_hr}
\title{Number of departures}
\usage{
departure_stop_line_hr(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{GTFS object loaded with tidytransit::read_gtfs()}
}
\value{
dep Table with number of departures per stop, hour and line for single date
}
\description{
Create dataframe with the number of departures per stop, hour and line for todays date
}
\examples{
\donttest{
library(dplyr)
gtfs_file <- list.files("gtfs_data")
dat = tidytransit::read_gtfs(paste0("gtfs_data/", gtfs_file[1]))
departure_stop_line_hr(dat)
}

}
