% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_download.R
\name{download_gtfs}
\alias{download_gtfs}
\title{Download GTFS zip file from trafiklab.se}
\usage{
download_gtfs(rkm)
}
\arguments{
\item{rkm}{Code for the regional public transport authority}
}
\description{
Download GTFS data and store zip file in created project folder. Requires key for Trafiklab GTFS Regional API
(https://www.trafiklab.se/api/trafiklab-apis/gtfs-regional/)
}
\examples{
\donttest{
library(rstudioapi)
library(httr)
download_gtfs(rkm = "ul")
}

}
