% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkm_code.R
\name{gtfs_rkm_kod}
\alias{gtfs_rkm_kod}
\title{Get codes for Swedish regional GTFS data}
\usage{
gtfs_rkm_kod()
}
\value{
df Table with public transport authority code (rkm) and corresponding regional name and ID. "rkm" serves as input parameter for gtfs_download()
}
\description{
Get codes for Swedish regional GTFS data
}
\examples{
\donttest{
gtfs_rkm_kod()
}

}
