% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{hpl_id_namn}
\alias{hpl_id_namn}
\title{Create dataframe with unique stop name and ID}
\usage{
hpl_id_namn(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{GTFS object loaded with tidytransit::read_gtfs()}
}
\value{
stops Table with unique stop ID and name
}
\description{
Create dataframe with unique stop name and ID
}
\examples{
\donttest{
library(dplyr)
gtfs_file <- list.files("gtfs_data")
dat = tidytransit::read_gtfs(paste0("gtfs_data/", gtfs_file[1]))
hpl_id_namn(dat)
}
}
