% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{lines_stop}
\alias{lines_stop}
\title{Lines per stop}
\usage{
lines_stop(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{GTFS object loaded with tidytransit::read_gtfs()}
}
\value{
lines Table with all lines per stop for single date
}
\description{
Create dataframe with all the lines per stop for todays date
}
\examples{
\donttest{
library(dplyr)
gtfs_file <- list.files("gtfs_data")
dat = tidytransit::read_gtfs(paste0("gtfs_data/", gtfs_file[1]))
lines_stop(dat)
}

}
