% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{n_lines_stop}
\alias{n_lines_stop}
\title{Number of lines per stop}
\usage{
n_lines_stop(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{GTFS object loaded with tidytransit::read_gtfs()}
}
\value{
lines Table with number of lines per stop for single date
}
\description{
Create dataframe with the number of lines per stop for todays date
}
\examples{
\donttest{
library(dplyr)
gtfs_file <- list.files("gtfs_data")
dat = tidytransit::read_gtfs(paste0("gtfs_data/", gtfs_file[1]))
n_lines_stop(dat)
}

}
