% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_sf.R
\name{network_sf}
\alias{network_sf}
\title{SF object with lines}
\usage{
network_sf(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{GTFS object loaded with tidytransit::read_gtfs()}
}
\value{
SF object with most common path for each line in WGS84
}
\description{
Create SF object with the public transport network from GTFS. Where a line has multiple different paths, the most common path is identified and included. So this is a simplified version of the network.
}
\examples{
\donttest{
library(dplyr)
library(sf)
library(mapview)
gtfs_file <- list.files("gtfs_data")
dat = tidytransit::read_gtfs(paste0("gtfs_data/", gtfs_file[1]))
network_sf(dat) \%>\% mapview()
}

}
