% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_linetype.R
\name{remove_linetype}
\alias{remove_linetype}
\title{Remove linetypes}
\usage{
remove_linetype(gtfs_obj, remove_linetype)
}
\arguments{
\item{gtfs_obj}{GTFS object loaded with tidytransit::read_gtfs()}

\item{remove_linetype}{String of line types to be removed}
}
\description{
Filter gtfs_obj by trips related to certain line types (route_desc), eg all trips, routes etc for "Stadsbuss"
}
\examples{
\donttest{
gtfs_file <- list.files("gtfs_data")
dat = tidytransit::read_gtfs(paste0("gtfs_data/", gtfs_file[1]))
remove_linetype(dat)
}

}
