% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{stop_id_name}
\alias{stop_id_name}
\title{Create dataframe with unique stop name and ID}
\usage{
stop_id_name(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{GTFS object loaded with tidytransit::read_gtfs()}
}
\value{
stops Table with unique stop ID and name
}
\description{
Create dataframe with unique stop name and ID
}
\examples{
\donttest{
library(dplyr)
gtfs_file <- list.files("gtfs_data")
dat = tidytransit::read_gtfs(paste0("gtfs_data/", gtfs_file[1]))
stop_id_name(dat)
}
}
