% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_sf.R
\name{stop_sf}
\alias{stop_sf}
\title{SF object with stops}
\usage{
stop_sf(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{GTFS object loaded with tidytransit::read_gtfs()}
}
\value{
SF object with stop coordinates in WGS84
}
\description{
Create SF object with stop coordinates from GTFS. Multiple stops with same ID are grouped where the position is based on the mean lat/long coordinates
}
\examples{
\donttest{
library(dplyr)
library(sf)
gtfs_file <- list.files("gtfs_data")
dat = tidytransit::read_gtfs(paste0("gtfs_data/", gtfs_file[1]))
stop_sf(dat)
}

}
